requirements.txt
Faker
phonenumbers
pyfiglet
pystyle
validators
cryptography
pycryptodome
